backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.0.2
postgrest==0.13.0
litellm==1.17.9
openai==1.3.7
requests==2.31.0
websockets==12.0
aioredis==2.0.1
celery==5.3.4
sentry-sdk[fastapi]==1.38.0
httpx==0.25.2
PyJWT==2.8.0
jwcrypto==1.5.0
email-validator==2.1.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
jsonschema==4.20.0
